requirements.txt
numpy==2.1.3
paho-mqtt==2.1.0
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
tzdata==2024.2
